backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.9.2
requests==2.32.3

